Token contract mapping, deposit recepient address
Token contracts on Ethereum will have corresponding contracts on the
DAppChain, the mapping can be specified in the genesis.

When an entity deposits tokens in the Gateway on Ethereum it must
specify the address of the entity that should ultimately receive the
tokens on the DAppChain.

The Solidity Gateway contract still needs to be updated (along with the
Go binding).

The token transfer test is still WIP too.