Handle nil value in Context.Set() consistently in internal/external plugins
The IAVLStore doesn't allow nil values, in internal/builtin plugins
(see wrappedPluginContext.Set()) no attempt will be made to store nil
values in the IAVLStore becase because proto.Marshal() will return an
empty byte array instead of nil if a protobuf message is encoded to nil
(because all its members are set to default values).

In the case of external plugins the relevant code is in
GRPCAPIServer.Set(), here the value encoded in the request message may
be nil, and the code will attempt to insert it into the IAVLStore cause
the IAVL tree to panic.