Don't call CheckTx if current block height is zero
This is meant to work around an issue that occurs when empty block
generation is switched off. The current block header is set in
Application.BeginBlock, with empty blocks switched off if a tx is sent
by a client to a node that has been restarted CheckTx will be called
before BeginBlock where the current block header is set, thus CheckTx
will not have access to imporant info like chain ID and block time.
To work around this issue we simply skip doing anything in CheckTx if
the block header is empty.