Register all contracts deployed via genesis & deploy txs
Registry V2 allows the Transfer Gateway to check who deployed a
particular contract on the DAppChain.

To maintain backwards compatibility with existing production DAppChains
the new registry will only be enabled if RegistryVersion is set to 2 in
loom.yml. There is no way for existing production DAppChains to switch
to the new registry as yet, we'll have to implement an upgrade process
that can be initiated via a tx.