Transfer Gateway ERC20 support (#282)
The DAppChain Gateway will only attempt to mint ERC20 tokens when it 
doesn't own enough of them to complete a transfer, so developers can
omit the `mintToGateway` function from their ERC20 contracts if they
wish to manually allocate tokens to the DAppChain Gateway. Alternatively
if the total supply is fixed then the whole amount can be allocated to
the DAppChain Gateway when the ERC20 contract is deployed.