Transfer Gateway ETH support (#290)
The DAppChain Gateway contract will store ETH that it receives from the
Mainnet Gateway contract in the ethcoin contract (an ERC20 like Go contract,
based on coin). Go contracts can then transfer the ETH around via the ethcoin contract.

ethcoin is just a copy of the coin contract, with an extra MintToGateway
method to allow the Gateway contract to mint ETH to complete transfers
from Mainnet to DAppChain.