Validators Tool Tweaks
Added the ability to specify a base config for nodes.

Also the contracts dir is now optional since you don't need one if
you're just running built-in contracts.

And the node RPC address is written to the node_rpc_addr file to make it
easier for clients to connect to nodes in the cluster.