Fix validators-tool to generate correct genesis for nodes
validators-tool has been generating invalid genesis files since it got
hacked to use DPOS v2 instead of v1, it only works with v2 now,
but the default in loom is still v1. So, to actually get validators-tool
to generates a cluster that starts you have to force nodes to be
initialized for DPOS v2, easy enough to do by providing a base loom.yaml
with `DPOSVersion: 2`. Unfortunately that didn't actually work because
the base loom.yaml for the cluster wasn't copied to the node directory
until after the node was initialized, these changes take care of that
issue.

Another issue is that the default genesis for the DPOS v2 contract had
an invalid Location field, which prevented the node from loading the
contract.