Set TM mempool wal dir
Supposedly the mempool wal is optional, so it now defaults to an empty
directory... except that doesn't prevent Tendermint from creating the
file, so not really optional... yet. This will probably get fixed in the
next TM release, but until then set the mempool wal dir to what it used
to default to in older TM versions.