Disable saving of failed EVM tx receipts & hashes
Previously if an EVM tx failed within `Application.CheckTx()` a tx
receipt would be retained, and saved to the app store in
`Application.EndBlock()`. A tx that errors out in `Application.CheckTx()`
will not be transmitted by the node to its peers, therefore the app hash
obtained by the node in `Application.Commit()` upon saving the current
block will not match the one obtained by its peers. At least in theory,
it's unclear whether we've seen this happen in practice, but I'm erring
on the side of caution here.

In the case of an EVM tx failing within `Application.DeliverTx()` the
tx receipt could be stored in the app state, as all nodes will do the
same. However, it doesn't seem like a good idea to modify app state
if a tx fails, even if it is technically possible at the moment, so
I've removed that code for the time being.

I think a better idea would be store tx receipts for failed txs in a
local cache that doesn't have to match across all validators, this cache
would store receipts for any EVM tx that failed, be it in `CheckTx`, or
`DeliverTx`.