Stop running tests in no-EVM builds
A lot of the tests are being run twice on Jenkins, once with the EVM
build and once with the no-EVM build. The more E2E tests we add the
slower this gets, since we're not actually shipping the no-EVM build
at the moment I'm disabling the tests on the no-EVM build. Tests will
still be built for the no-EVM build to ensure that build mode remains
functional should we actually need it, can switch them back on if we
rearrange the build pipeline a bit to run the E2E tests separately.