Add `abci-server` flag to `loom run` command
The abci-server flag can be used to specify a unix socket address which
will be used to serve the ABCI app. When this flag is specified the node
won't start any of the Tendermint services (aside from the one that
serves the ABCI app), it's expected that an external Tendermint process
will drive the ABCI app.