Stop running tests in no-EVM builds
A lot of the tests are being run twice on Jenkins, once with the EVM
build and once with the no-EVM build. Some of the e2e tests may not
work in the no-evm builds anymore, so I'm disabling the no-evm test run
until we have more time to investigate.