Rename AppendSourcesForUser to AddKarma
AppendSourcesForUser did indeed append new karma sources, but it also
overwrote existing source counts, so the name was a bit inaccurate.
Renamed the method for clarity, and stopped it from overwriting existing
source counts, which it shouldn't have been doing.

Renamed validatedUpdateSourcesForUser to setKarmaForUser, and simplified
it.

Replaced KarmaStateUser type with AddKarmaRequest, and
KarmaAddressSource.