Add --ext flag to debug dump-mempool
When the -e / --ext flag is specified the mempool_txs endpoint will be
queried instead of unconfirmed_txs, mempool_txs provides additional
metadata for each tx (currently just the height at which each tx was
added to the mempool). When --ext is specified the --limit flag can be
set to -1 to fetch all txs in the mempool.