Check minimum build version before enabling feature flag (#982)
Force the ChainConfigManager to panic if it tries to enable a feature that's not supported by the current build. This should safely prevent the node from processing any further blocks in case a feature it doesn't support is activated by the rest of the chain. To make this work features now have a minimum build number associated with them, which must be specified by the ChainConfig contract owner when adding new features.

The new functionality is not behind a feature flag, so once this is deployed to production new features can't be added until the whole cluster upgrades to a new build.