Fix: delegator gets extra funds when redelegating
When delegations that are getting redelegated get processed in distributeDelegatorRewards during the next election, the original delegation is essentially refunded to the delegator, and a new delegation with req.Amount goes into effect.  So the net effect is that the delegator receives req.Amount in new money.

Fixed by setting the state during redelegation to REDELEGATING, and updating the updateValidator to be the same as before