Authorization Logic for ConfirmWithdrawalReceiptV2 (#977)
In this PR, we define two authorization methods for ConfirmWithdrawalReceiptV2. The TRUSTED_VALIDATORS and the DPOS_VALIDATORS.

When TRUSTED_VALIDATORS is toggled, the signatures submitted in ConfirmWithdrawalReceiptV2 are checked against the Ethereum mapping of validatorsAuthConfig.TrustedValidators.Validators, and if they are all correctly recovered, the call is authorized. 

When DPOS_VALIDATORS is toggled, the signatures are checked against the ethereum mapping of ctx.Validators(), and if more than 2/3rds of the stake signed, then the call is authorized.

For now, we'll use the TRUSTED_VALIDATORS config, and the DPOS_VALIDATORS will be enabled only after we add slashing and other better security mechanisms. The strategy can be changed by the gateway owner

OracleSignature which is a serialized bytes array of length 65 bytes (each one being a signature). The previous authorization method of ConfirmWithdrawalReceipt involved making a check against just a OracleAddress.