Don't enable the caching store for the multi writer app store
The caching store doesn't work with the multi writer app store
snapshots, so the snapshots bypass the cache, which means the QueryServer
never hits the cache in the caching store, no point writing to the cache
if nothing ever reads from it.

Also simplified the Range implementation of the multi writer app store
and the EVM store, we don't really use Range(nil) anywhere, and there
was a lot of code to support that use case that just added a lot of
additional complexity.