Add TRON Gateway contract & TG Oracle
The Gateway Go contract has been extended to support deposit & withdrawal
of TRX & TRC20 tokens. A new instance of the contract will be need to be deployed
under the name "tron-gateway" to handle all TRX & TRC20 tokens on the DAppChain.

TRX needs to be mapped to an EVM ERC20 contract on the DAppChain before deposits
& withdrawals can be made, this mapping can only be made by the "tron-gateway"
contract owner.

The TG Oracle has been extended to support fetching deposit & withdrawal events
from TRON. A new instance of the TG Oracle will be launched to bridge TRON to
the "tron-gateway" contract on the DAppChain.