Improve EvmStore.getLastSavedRoot performance (#1159)
The way we called ReverseIterator was not quite right. ReverseIterator takes two arguments start and end. TM ReverseIterator uses seek(end) to go to the latest prefix key and then iterate back to start. Previously, we only specified start which made it iterate through the database from the last index. This has now been fixed by specifying the end of the range for the reverse iterator.