Only write EVM state root to IAVL store when the root changes (#1214)
Previously MultiWriterAppStore wrote the EVM root to app.db on every block, this seems to have caused some kind of rebalancing of the IAVL tree, which resulted in the app hash changing every block, which in turn forced Tendermint to create a new empty block. As a result, `CreateEmptyBlocks: false` in `loom.yaml` stopped working as intended when the MultiWriterAppStore was enabled. To fix this problem we now check if the EVM root has actually changed before writing it to the IAVL store. A new feature flag has been added to deploy this fix to production.