Remove OriginValidator
The rate limiting the OriginValidator was supposed to do belongs in
middleware, and the deployer whitelist is now handled by the
DeployerWhitelist contract & middleware. The CallTx throttling in the
OriginHandler is also a bit awkward. There little point in keeping the
OriginValidator around any longer in its current form since it was never
activated in production, so I'm removing it to make way for another tx
throttling middleware.