Remove ThrottleTxMiddleware & OriginValidator (#1233)
* Remove ThrottleTxMiddleware

This middleware has been superseded by the DeployerWhitelist contract
and corresponding middleware, the old middleware is still needed to
replay some very old PlasmaChain blocks - but since it's no longer
practical to replay blocks from that far back I'm getting rid of this
old middleware.

* Remove OriginValidator

The rate limiting the OriginValidator was supposed to do belongs in
middleware, and the deployer whitelist is now handled by the
DeployerWhitelist contract & middleware. The CallTx throttling in the
OriginHandler is also a bit awkward. There little point in keeping the
OriginValidator around any longer in its current form since it was never
activated in production, so I'm removing it to make way for another tx
throttling middleware.