Return tx hashes for all transactions & receipts for all valid hashes. (#1226)
Addresses #1157
Should also provide a fix for #1190 eth_getTransactionReceipt

Allows accessing receipts using both Tendermint and loom tx hashes.
In block, inquiries will display tendermint tx hash if no loom tx hash is available, for example, if the transaction failed.

Will return a receipt if the input is a valid hash (starts 0x) and the loomchian is ok. If the hash cannot be identified as either a loom-tx-hash or tendermint-tx-hash, an empty receipt with status ~ failed is returned.