Wait the offline node to start before unjailing it (#1286)
Jailing offline validator e2e test is still flaky. I am suspecting that the offline node has not started and downtime periods are not cleared out before we unjail it. As a result, the node is jailed again immediately after it is unjailed.

The test depends on Delay which is not reliable. This PR implements wait_node_to_start command that makes the test wait until the offline node has started and downtime periods have been cleared out before unjailing the offline node.