Put user deployer whitelist contract changes behind a feature flag
The new fields added to support per-contract tx limiter cannot be set
until the entire cluster has been upgraded.

Also fix e2e test to detect errors & fix divide by zero in tx limiter.
The block range for a tier should be non-zero, but if someone forgets to
set it on an existing cluster where a previous version of the user
deployer whitelist contract was deployed the node shouldn't panic.