Ensure slashValidatorDelegations doesn't cause consensus failure
slashValidatorDelegations is called by UnregisterCandidate so it can't
modify contract state until the downtime slashing feature flag is
enabled.

slashValidatorDelegations may also be called by rewardAndSlash on each
election, but in practice that won't happen until the downtime slashing
feature flag is enabled because the slash percentage on each static will
be zero until that feature is enabled.