Fix flaky Karma Coin Test (#1423)
Karma Coin test is still flaky and we still cannot identify why it fails. According to the log, the approve request has been processed successfully but the allowance amount is still 0.

So I suspect that FakeContextWithState might be the culprit as it is only used by this Karma Coin test. This FakeContextWithState is unnecessary and not used by any other tests.

This PR replaces FakeContextWithState with FakeContext in Karma Coin test.