Add on-chain config (#1277)
The on-chain config provides a way to tweak cluster-wide settings, 

Each change to the on-chain config has an associated minimum build number, once a sufficient number of the validators have upgraded to that build the change is applied to the on-chain config. Any node that isn't running the minimum required build when a config change goes into effect will panic in `Application.BeginBlock`, in most cases it should be possible to bring the node back online by simply restarting the node on a newer build.

The new on-chain config functionality will only be available once the `chaincfg:v1.3` feature flag is activated. The loom CLI has been extended with the following commands:
- chain-cfg list-pending-actions - shows a list of pending changes to the on-chain config.
- chain-cfg config - shows the current on-chain config.
- `chain-cfg set-setting <key> --value <string> --build <min-build-number> -k <chain-cfg-owner-key>`

Only the ChainConfig contract owner can change on-chain config settings.