Disable saving of EVM tx receipts for EVM calls from Go contracts by default (#1473)
PR #1401 introduced a replay breaking change that should've been behind a feature flag. The breakage could be reproduced whenever a block included a Go tx that called the EVM, followed by one or more EVM txs. The hashes generated for EVM txs currently depend on the number of EVM tx receipts saved in each block because the EVM tx hash is currently generated by hashing the EVM tx receipt, which includes the `TransactionIndex`. The `TransactionIndex` in the receipt is currently set to match the number of EVM tx receipts generated in the block thus far, so saving one extra receipt ends up changing the EVM tx hashes in the block. The EVM tx hashes are included in the DeliverTx result which Tendermint hashes as part of the block info.

To address this problem a new `receipts:v3` feature flag has been added to safely enable the new behaviour on clusters where the build with the breaking change hasn't been deployed yet. On clusters where this behaviour has already been enabled a `ReceiptsVersion: 3` must be set in `loom.yml` when upgrading to the new build, followed by enabling of the `receipts:v3` feature flag. The `loom.yml` setting will be deprecated in the near future once all clusters have the new feature flag enabled.