Store EVM tx receipts for txs that fail (#1399)
Previously EVM tx receipts were only stored for EVM txs that didn't error out, this presented a problem because EVM txs that fail are still committed to a block and the block explorer expects to be able to find a receipt for each tx in a block. Now (after the `receipts:v3` feature flag is activated) receipts will be stored for all EVM txs, regardless of whether they succeeded or not.

NOTE: `NonceHandler.IncNonceOnFailedTx` should be enabled in the on-chain config before the `receipts:v3` feature flag. This will ensure that EVM tx hashes are generated with the correct nonce.

Co-authored-by: Vadim Macagon <vadim.macagon@gmail.com>