Fix EVM tx hash generation algorithm (#1481)
Currently the tx hash generated for an EVM tx is actually a hash of the tx receipt (not even the final one, some intermediate version of it), this makes it impossible for a client to figure out what the tx hash will be before submitting an EVM tx. When the `receipts:v2` feature flag is enabled the EVM transaction hash generation algorithm will change to mimic Ethereum so that the hash can be obtained client-side before the tx is submitted to a node.