Fix missing version info in TG Oracles
The git SHAs of the various packages used to build the Oracle binaries
are supposed to be filled in via build flags, the flags weren't set
correctly in the loomchain Makefile so the Oracle binaries were being
built with no versioning info. The Oracle targets in the loomchain
Makefile now delegate the actual builds to the transfer-gateway Makefile
so that the versioning info is embedded in the Oracle binaries as
intended.