Ensure tx hash on returned tx receipt always matches the requested one (#1513)
When a Go contract calls an EVM contract a tx receipt is generated, that receipt contains the tx hash generated for the EVM call (this is essentially an internal call). Later on Blockscout fetches the block containing the Go contract -> EVM contract CallTx and attempts to fetch the receipt by using the Tendermint hash of the SignedTx bytes. The Tendermint tx hash is resolved to an EVM tx hash, and the corresponding receipt is returned. Previously the returned receipt contained the EVM tx hash rather than the Tendermint tx hash that Blockscout supplied in the request, which confused Blockscout. Now before the receipt is returned to the client the tx hash on the receipt will always be set to match the one that was requested.