Ensure eth_getLogs returns tx hashes that are consistent with eth_getBlockByNumber (#1514)
The tx hashes in the logs returned by eth_getLogs don't always match the ones returned by eth_getBlockByNumber for the same block. The mismatch occurs for txs that call a Go contract that in turn calls an EVM contract, the receipt for the EVM call is stored with the EVM tx hash as the key, and that EVM tx hash is then returned in the logs. This discrepancy between eth_getLogs and eth_getBlockByNumber confuses Blockscout. These changes fix that discrepancy.