Make chain-cfg list-validators display only active validators by default (#1530)
Previously this command showed the build number for all validators known to the chain config contract, this included candidates that were not part of the active validator set (because they've been voted out). These changes make the command output only the info for the active validator set by default, the new `--all` flag can be used to display all known validators (active or not).