Make node panic on unsupported build number (#1507)
Previously it was possible to start a node from a jump-start on an older build than the one that was used to create the jump-start. If there were no hard fork changes this worked fine, but if the newer build was used to activate features or apply on-chain config changes that weren't supported by the older build the node was highly likely to encounter an app hash mismatch sooner or later.

To prevent such mishaps the minimum required build is now tracked on-chain, so each time a feature flag is activated, or an on-chain config setting is applied the minimum required build is bumped if needed. The node checks the minimum build number on startup and panics if it detects the current build is older than the minimum required build at the node's current app height.