Fix nonce handler truffle tests
These tests were using eth_getTransactionCount to obtain the account
nonce, this method now expects to be given an Ethereum address instead
of a DAppChain address. To fix this an account mapping is now created
prior to the each nonce test case and the Ethereum address is used in
calls to eth_getTransactionCount.