Make eth_sendRawTransaction accept signed Ethereum txs (#1227)
The eth_sendRawTransaction RPC method on the /eth endpoint now works similarly to the way one would expect it to work on an Ethereum node. Any Ethereum tx submitted via the eth_sendRawTransaction RPC method is first wrapped in a SignedTx, which is then submitted to the mempool. The SignedTx itself is unsigned so the signer address is recovered from the wrapped Ethereum tx and mapped to a local account on the chain, which means that an account mapping between the signer's Ethereum account and a local account must exist before eth_sendRawTransaction can accept signed Ethereum txs from that signer.

A number of other issues with the Web3 JSON-RPC interface have been fixed (some related, some not so much):
- https://github.com/loomnetwork/loomchain/issues/1377
- https://github.com/loomnetwork/loomchain/issues/1376
- https://github.com/loomnetwork/loomchain/issues/1434

eth_getTransactionCount has also been fixed, it now expects to be given an Ethereum address instead of a local address, a mapping must exist between the Ethereum address and a local address for this method to return a valid nonce. The nonce handler truffle tests have been fixed to call eth_getTransactionCount with an Ethereum address instead of a local address, and account mapping are now created prior to the each nonce test case.

Co-authored-by: Vadim Macagon <vadim.macagon@gmail.com>