Simplify setup of gateway reactor non-validator nodes (#1543)
These changes add a new `FnConsensus.Reactor.IsValidator` setting to loom.yml:

FnConsensus:
  Reactor:
    IsValidator: false

This setting can be used to run a node that forwards messages sent via the custom reactor without signing any of them, and doing only basic validation before broadcasting the messages to its peers. This is the mode in which sentry nodes are expected to run.