Adjust tx nonce in eth_sendRawTransaction (#1550)
The following fixes address a set of issues that were encountered while using the Truffle HDWallet provider to deploy & test Solidity contracts to a test cluster.

* Propagate errors properly so Truffle shows more than just "Server error"

* eth_getBlockByNumber should return null for blocks that don't exist

In our case that also includes the "pending" block. If anything else is
returned for a non-existent block Truffle (or the provider) gets very
confused.

* Adjust nonces on txs send to eth_sendRawTransaction

Loom tx nonces start at 1, Ethereum tx nonces start at zero, when an
Ethereum tx is sent to eth_sendRawTransaction the nonce needs to be
adjusted to account for this difference.

* Fix a few of the test to work with the Truffle HDWallet provider

Mainly just need to account for the fact that the accounts provided by
the Truffle HDWallet provider have Ethereum addresses while the ones
provided by Loom Truffle provider are Loom addresses.

* Run Truffle tests using Loom Truffle provider & HDWallet provider

Only a subset of the Truffle tests run with the HDWallet provider, the
ones that don't really use the provider just run once with the Loom
Truffle provider.