Emit tx events after block commit (#1562)
Previously tx events were emitted to subscribers immediately after the tx was processed in DeliverTx, however at that point the state changes from the tx haven't been persisted to disk yet. If the subscriber attempted to query the contract state as soon as they received the tx event it's possible they wouldn't see the state changes from that tx. To reduce the likelihood of subscribers "missing" state changes in the aforementioned scenario the tx events are now emitted only after the state changes have been written out to disk.