Eliminate PruningIAVLStore
We haven't used this store in ages and have no intention of doing so
again because it has been replaced by the IAVL flushing interval.
Finally getting rid of this store now because the snapshot code for the
remaining stores needs to be updated and I don't want to waste time
modifying code that doesn't get used.