Revert some modifications to the MultiWriterAppStore tests
The tests are supposed to test how the vm-prefixed keys are handled,
replacing the prefix makes the tests fairly pointless, so the prefix
has been restored.

A couple of the snapshot tests have been removed because the EvmStore
no longer caches writes until Commit, it now writes them out immediately.