Hack VersionedCachingStore to get the e2e tests passing
The EVM root in the VersionedCachingStore cache wasn't being updated in
VersionedCachingStore.SaveVersion() which meant that the
snapshots obtained from the VersionedCachingStore contained a stale EVM
root that didn't match the EVM state in the snapshot.

This hack updates the cached EVM root in
VersionedCachingStore.SaveVersion() but it's not a proper fix for the
issue, the VersionedCachingStore shouldn't be aware of the EVM root at
all. The most sensible thing to do is probably to eliminate the
MultiWriterAppStore itself and update Application.Commit() to call
EvmStore.Commit(), and then store the EVM root in the
VersionedCachingStore.