Refactor the way the on-chain flush interval override is applied
There was a bug in EvmStore.Commit() where it attempted to load the
on-chain config from evm.db, but that's not where the on-chain config
gets persisted to - it's persisted to app.db. This meant that the
flush interval specified in the on-chain config did not override the
one specified in the EvmStore section in loom.yml as was originally
intended.