Add ability to copy out the currently used state from the app DB to a new DB (#1705)
To copy the currently used state at a particular height the
`AppStore.CloneStateAtHeight` setting must be set in the on-chain config to
the app height / block number at which the copy should be made. When the
app reaches that height it will copy out the current state from `app.db` to
`app_v2.db` (omitting ancient keys that are no longer used), from that point
on all further state changes will (i.e. IAVL tree versions) will be saved to the
new DB.

When the node starts up it will now check whether `app.db` or `app_v2.db`
contains the most recent state (latest IAVL tree version), if it detects that
`app_v2.db` has the most recent state it will replace `app.db` with `app_v2.db`,
like so:
- Rename `app.db` -> `old_app_timestamp.db`
- Rename `app_v2.db` -> `app.db`
It will then proceed to use `app.db` as usual.

A number of `db` sub-commands have been added to the CLI for offline
analysis of `app.db` and testing & verification of the state copying/cloning.

Other changes:
* Fix input redirection in the kill_and_restart_node e2e test command

The command didn't seem to redirect the output after restarting a node,
which made it impossible to figure out why a node failed to restart.

* Don't override the DPOSv3 InitCandidates setting in validators-tool

Not sure why this was hardcoded to false before, but now it'll be read
from the genesis as it should've been.

* Switch `iavl` package to `loomchain2` branch