Support gas parameter when calling eth_estimateGas
eth_estimateGas accepts a `gas` parameter that can be used to limit the
maximum amount of gas that can be used during estimation. Note that the
gas limit that's provided is capped at the per-tx gas limit set via the
on-chain config.

Also fixed some overly lax validation of from/to address parameters,
while it's acceptable to omit them in some scenarios it's not acceptable
to ignore invalid address values.